Homework 4 1. Создал xhtml страницы: каталог, корзины, оформления заказа, страница о компании. Использована технология JSF. 2. Для реализации JSF переделаны: controller, listener, persist; 3. Добавлено редактирование, добавление, удаление продукта.
