+++ store-jsf/src/main/java/ru/geekbrains/persist/Product.java
