+++ store-jsf/src/main/webapp/WEB-INF/views/head.xhtml
@@ -1,0 +1,46 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:h="http://xmlns.jcp.org/jsf/html"
+      xmlns:ui="http://xmlns.jcp.org/jsf/facelets"
+      xmlns:f="http://xmlns.jcp.org/jsf/core">
+
+<h:head>
+    <!-- Required meta tags -->
+    <meta charset="utf-8"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no"/>
+
+    <!-- Bootstrap CSS -->
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+          integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous"/>
+    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.3.1/css/all.css"/>
+    <title>#{title}</title>
+</h:head>
+<h:body>
+<nav class="navbar navbar-expand-lg navbar-light bg-light">
+
+    <ul class="navbar-nav mr-auto">
+        <li class="nav-item active">
+            <a class="nav-link" href="index.xhtml">Каталог</a>
+        </li>
+
+        <li class="nav-item active">
+            <a class="nav-link" href="cart.xhtml">Корзина</a>
+        </li>
+
+        <li class="nav-item active">
+            <a class="nav-link" href="order.xhtml">Корзина</a>
+        </li>
+
+        <li class="nav-item active">
+            <a class="nav-link" href="about.xhtml">О компании</a>
+        </li>
+    </ul>
+
+</nav>
+<div class="container">
+<div class="row">
+    <h2>#{name}</h2>
+</div>
+</div>
+</h:body>
+</html>
