+++ store-jsf/src/main/webapp/WEB-INF/web.xml
@@ -6,13 +6,14 @@
   <display-name>Archetype Created Web Application</display-name>
 
   <servlet>
-    <servlet-name>Main</servlet-name>
-    <servlet-class>ru.geekbrains.Main</servlet-class>
+    <servlet-name>faces-servlet</servlet-name>
+    <servlet-class>javax.faces.webapp.FacesServlet</servlet-class>
+    <load-on-startup>1</load-on-startup>
   </servlet>
 
   <servlet-mapping>
-    <servlet-name>Main</servlet-name>
-    <url-pattern>/</url-pattern>
+    <servlet-name>faces-servlet</servlet-name>
+    <url-pattern>*.xhtml</url-pattern>
   </servlet-mapping>
 
 
@@ -23,6 +24,12 @@
     <url-pattern>*.js</url-pattern>
     <url-pattern>*.jpeg</url-pattern>
   </servlet-mapping>
+
+<welcome-file-list>
+  <welcome-file>index.jsp</welcome-file>
+  <welcome-file>index.html</welcome-file>
+  <welcome-file>index.xhtml</welcome-file>
+</welcome-file-list>
 
   <context-param>
     <param-name>jdbcConnectionString</param-name>
@@ -38,8 +45,6 @@
     <param-name>password</param-name>
     <param-value>123123</param-value>
   </context-param>
-
-
 
 </web-app>
