+++ store-jsf/src/main/webapp/about.xhtml
@@ -1,0 +1,27 @@
+<html lang="en"
+      xmlns:h="http://java.sun.com/jsf/html"
+      xmlns:f="http://xmlns.jcp.org/jsf/core"
+      xmlns:ui="http://java.sun.com/jsf/facelets">
+
+<ui:include src="WEB-INF/views/head.xhtml">
+    <ui:param name="title" value="About"/>
+    <ui:param name="name" value="О компании"/>
+</ui:include>
+
+
+<h:body>
+    <div class="container">
+    <div class="row">
+        Повседневная практика показывает, что укрепление и развитие структуры обеспечивает широкому кругу (специалистов)
+        участие в формировании дальнейших направлений развития. Товарищи! консультация с широким активом позволяет выполнять
+        важные задания по разработке систем массового участия. Таким образом реализация намеченных плановых заданий
+        позволяет оценить значение новых предложений. Равным образом постоянный количественный рост и сфера нашей
+        активности играет важную роль в формировании системы обучения кадров, соответствует насущным потребностям.
+        Равным образом рамки и место обучения кадров влечет за собой процесс внедрения и модернизации системы обучения
+        кадров, соответствует насущным потребностям. Идейные соображения высшего порядка, а также рамки и место обучения
+        кадров обеспечивает широкому кругу (специалистов) участие в формировании новых предложений.
+    </div>
+    </div>
+
+</h:body>
+</html>
