+++ store-jsf/src/main/webapp/cart.xhtml
@@ -1,0 +1,21 @@
+<html lang="en"
+      xmlns:h="http://java.sun.com/jsf/html"
+      xmlns:f="http://xmlns.jcp.org/jsf/core"
+      xmlns:ui="http://java.sun.com/jsf/facelets">
+
+<ui:include src="WEB-INF/views/head.xhtml">
+    <ui:param name="title" value="Cart"/>
+    <ui:param name="name" value="Корзина"/>
+</ui:include>
+
+
+<h:body>
+    <div class="container">
+        <div class="row">
+            С другой стороны укрепление и развитие структуры обеспечивает участие в формировании систем массового участия.
+            Идейные соображения высшего порядка, а также дальнейшее развитие различных форм деятельности позволяет оценить значение новых предложений.
+        </div>
+    </div>
+
+</h:body>
+</html>
