+++ store-jsf/src/main/webapp/index.xhtml
@@ -1,0 +1,78 @@
+<html lang="en"
+      xmlns:h="http://java.sun.com/jsf/html"
+      xmlns:f="http://xmlns.jcp.org/jsf/core"
+      xmlns:ui="http://java.sun.com/jsf/facelets">
+
+<ui:include src="WEB-INF/views/head.xhtml">
+    <ui:param name="title" value="Catalog"/>
+    <ui:param name="name" value="Каталог"/>
+</ui:include>
+
+
+<h:body>
+
+    <div class="container">
+        <div class="row py-2">
+
+            <h:form id="productForm">
+
+
+                <div class="col-12">
+
+                    <h:dataTable value="#{productController.allProducts}" var="product"
+                                 styleClass="table table-bordered my-2">
+
+                        <h:column>
+                            <f:facet name="header">Id</f:facet>
+                            #{product.id}
+                        </h:column>
+
+                        <h:column>
+                            <f:facet name="header">Title</f:facet>
+                            #{product.title}
+                        </h:column>
+
+                        <h:column>
+                            <f:facet name="header">Description</f:facet>
+                            #{product.description}
+                        </h:column>
+
+                        <h:column>
+                            <f:facet name="header">Price</f:facet>
+                            #{product.price}
+                        </h:column>
+
+                        <h:column>
+                            <f:facet name="header">Actions</f:facet>
+                            <h:commandButton value="Edit" styleClass="btn btn-success"
+                                             action="#{productController.editProduct(product)}"/>
+                            <h:commandButton value="Delete" styleClass="btn btn-danger"
+                                             action="#{productController.deleteProduct(product)}">
+                                <f:ajax event="click" render="productForm"/>
+                            </h:commandButton>
+                        </h:column>
+                    </h:dataTable>
+                </div>
+
+                <div class="col-12">
+                    <h:commandButton value="Add Product" styleClass="btn btn-primary"
+                                     action="#{productController.createProduct()}"/>
+                </div>
+
+            </h:form>
+        </div>
+    </div>
+    <!-- Optional JavaScript -->
+    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
+    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
+            integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+            crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
+            integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
+            crossorigin="anonymous"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
+            integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
+            crossorigin="anonymous"></script>
+
+</h:body>
+</html>
