+++ store-jsf/src/main/webapp/order.xhtml
@@ -1,0 +1,22 @@
+<html lang="en"
+      xmlns:h="http://java.sun.com/jsf/html"
+      xmlns:f="http://xmlns.jcp.org/jsf/core"
+      xmlns:ui="http://java.sun.com/jsf/facelets">
+
+<ui:include src="WEB-INF/views/head.xhtml">
+    <ui:param name="title" value="Order"/>
+    <ui:param name="name" value="Заказ"/>
+</ui:include>
+
+
+<h:body>
+    <div class="container">
+        <div class="row">
+            Равным образом рамки и место обучения кадров влечет за собой процесс внедрения и модернизации системы обучения
+            кадров, соответствует насущным потребностям. Идейные соображения высшего порядка, а также рамки и место обучения
+            кадров обеспечивает широкому кругу (специалистов) участие в формировании новых предложений.
+        </div>
+    </div>
+
+</h:body>
+</html>
