+++ store-jsf/src/main/webapp/product.xhtml
@@ -1,0 +1,62 @@
+<html lang="en"
+      xmlns:h="http://java.sun.com/jsf/html"
+      xmlns:p="http://xmlns.jcp.org/jsf/passthrough"
+      xmlns:f="http://xmlns.jcp.org/jsf/core"
+      xmlns:ui="http://java.sun.com/jsf/facelets">
+
+
+<ui:include src="WEB-INF/views/head.xhtml">
+    <ui:param name="title" value="Product Edit"/>
+    <ui:param name="name" value="Редактирование и добавление продукта"/>
+</ui:include>
+<h:body>
+
+
+    <div class="container">
+        <div class="row py-2">
+            <div class="col-12">
+                <h:form>
+                    <input type="hidden" id="id" name="id"/>
+                    <div class="form-group">
+
+                        <h:outputLabel value="Title"/>
+                        <h:inputText styleClass="form-control" id="title"
+                                     value="#{productController.product.title}"
+                                     placeholder="Enter title"/>
+                    </div>
+                    <div class="form-group">
+
+                        <h:outputLabel value="Description"/>
+                        <h:inputText styleClass="form-control" id="Description"
+                                     value="#{productController.product.description}"
+                                     placeholder="Enter description"/>
+                    </div>
+
+                    <div class="form-group">
+
+                        <h:outputLabel value="Price"/>
+                        <h:inputText p:type="price" styleClass="form-control" id="Price"
+                                     value="#{productController.product.price}"
+                                     placeholder="Enter price"/>
+                    </div>
+
+                    <h:commandButton value="Submit" styleClass="btn btn-primary"
+                                     action="#{productController.saveProduct()}"/>
+                </h:form>
+            </div>
+        </div>
+    </div>
+
+    <!-- Optional JavaScript -->
+    <!-- jQuery first, then Popper.js, then Bootstrap JS -->
+    <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
+            integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+            crossorigin="anonymous"></script>
+    <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
+            integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
+            crossorigin="anonymous"></script>
+    <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
+            integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
+            crossorigin="anonymous"></script>
+</h:body>
+</html>
